requirements.txt
psycopg2-binary
python-dotenv
django 
djangorestframework
djangorestframework-simplejwt
ipdb